Move files and add class bgTasks
